requirements.txt
Pillow==2.5.3
numpy==1.9.0
